测试用例支持数据类型扩展

+++ mxrec_add_ons/rec_for_torch/operators/jagged_to_padded_dense/op_host/jagged_to_padded_dense.cpp
@@ -122,7 +122,7 @@ public:
     {
         this->Input("values")
             .ParamType(REQUIRED)
-            .DataTypeList({ge::DT_FLOAT, ge::DT_INT64, ge::DT_FLOAT16, ge::DT_BF16, ge::DT_INT64})
+            .DataTypeList({ge::DT_FLOAT, ge::DT_INT64, ge::DT_FLOAT16, ge::DT_BF16, ge::DT_INT32})
             .FormatList({ge::FORMAT_ND});
         this->Input("offsets")
             .ParamType(REQUIRED)
